add back codes for downloading from http source (#278)
Summary:
with Django's file storage logic, the client side would build a binary and upload it to the Django server, with the file path identified with an http address. The lab side would then download the file from the path. Thus, these lines are necessary for the lab site to properly download the file from the server.
Pull Request resolved: https://github.com/facebook/FAI-PEP/pull/278

Reviewed By: sf-wind

Differential Revision: D15180613

Pulled By: hl475

fbshipit-source-id: 9eddc6eda8a8dd4a4b5885e529d00ad82232d7eb