Remove apps after benchmark in ios tests
Summary:
Sometimes the benchmark may fail due to some compatibility issue. It is unclear what the problem is.
At this moment, we have found that the installed app before the benchmark may impact whether the
benchmark is successful or not. However, we always remove the previous app before installing the
new one. But, to be on the safe side, we remove the apps after each benchmark to avoid this confusion.

Reviewed By: hl475

Differential Revision: D15437181

fbshipit-source-id: c09f5b5a6e756263150f3445106de9e63ce1c38c