allow legacy codes (#331)
Summary:
Pull Request resolved: https://github.com/facebook/FAI-PEP/pull/331

After D16676619, we found there are some FAILED and USER_ERROR from the run, either due to the legacy way of writing json or the json is actually testing multi-thread.

If we say `"num_threads": -1` in `test` filed, then we will do nothing. This behavior is similar to `"iter": -1. - tech debt

Again, should only apply to `host` case.

Reviewed By: BIT-silence

Differential Revision: D16723763

fbshipit-source-id: a3bffd2ee7c7eb9064ba85ff5862f4a52f297224