Add async mode to processRun (#335)
Summary:
Pull Request resolved: https://github.com/facebook/FAI-PEP/pull/335

processRun has been refactored into processRun and processWait. If async=True is passed into processRun then a tuple of the Popen object and the timeout will be returned. In this case it is the responsibility of the caller to pass this tuple into processWait in order to gather the output from the process.

Reviewed By: hl475

Differential Revision: D16997805

fbshipit-source-id: 5a615eaacb79dae4b414c7d96323bb5ebb7c583d