Allow config overrides to be read in from config_overrides.txt (#341)
Summary:
Pull Request resolved: https://github.com/facebook/FAI-PEP/pull/341

Currently we have a different set of cli args for every machine. Since we want
to start aibench using a chef config we need the command for starting aibench
to be the same for all machines. In order to do that we are allowing machine
specific configs to be read from config_overrides.txt instead of from cli
flags.

Reviewed By: hl475

Differential Revision: D17216679

fbshipit-source-id: f8c7b96defa80d1c960a1e4eb923ee4d942c8a07