Enable storage and retreival of multiple run statuses
Summary:
run status was stored in a global variable which allowed for race conditions
when multiple processes were started concurrently and causes us to lose the
return status from one of the processes. We now associate run status's with a
key. By default the key is an empty string which is used for the benchmark
command. Other processes that started concurrently should use their own key.

Differential Revision: D17557861

fbshipit-source-id: 60bdde93ebd9e5f322ffff70c8890e7a23edb187