Enable storage and retreival of multiple run statuses (#360)
Summary:
Pull Request resolved: https://github.com/facebook/FAI-PEP/pull/360

run status was stored in a global variable which allowed for race conditions
when multiple processes were started concurrently and causes us to lose the
return status from one of the processes. We now associate run status's with a
key. By default the key is an empty string which is used for the benchmark
command. Other processes that started concurrently should use their own key.

Reviewed By: hl475

Differential Revision: D17557861

fbshipit-source-id: c782eba490611bbdc1fb5e3cb35eed9a530a0889