Fix bug with forking test binaries
Summary:
When benchmarks run under bash they fork additional child processes that must
be killed in order to the benchmark to be timed out. We now start each
subprocess in it's own process group. So all of the child processes of the
benchmark inherit that process group. Then when the benchmark times out we can
kill the entire process group.

Differential Revision: D17726423

fbshipit-source-id: e172592206706f35c878d7412083671cb52fe9af