Fix bug with forking test binaries (#367)
Summary:
Pull Request resolved: https://github.com/facebook/FAI-PEP/pull/367

When benchmarks run under bash they fork additional child processes that must
be killed in order to the benchmark to be timed out. We now start each
subprocess in it's own process group. So all of the child processes of the
benchmark inherit that process group. Then when the benchmark times out we can
kill the entire process group.

Reviewed By: hl475

Differential Revision: D17726423

fbshipit-source-id: 3daf0c0183e1f3cf4f9141ae9be82dab725fa137