Temporary commit at 10/15/2019, 9:25:35 PM
Summary:
This is a change in order to consolidate D17922360

Our assumption is control is always a committed diff, so we should have diff_number and commit_hash. For treatment, it should have diff_number and commit_hash if it is a post_detection, and diff_number and diff_version.

Differential Revision: D17948428

fbshipit-source-id: 011f513288330338361be5a35bd7d950afaed2f3