update meta (#376)
Summary:
Pull Request resolved: https://github.com/facebook/FAI-PEP/pull/376

This is a change in order to consolidate D17922360

Our assumption is control is always a committed diff, so we should have diff_number and commit_hash. For treatment, it should have diff_number and commit_hash if it is a post_detection, and diff_number and diff_version.

Reviewed By: houseroad

Differential Revision: D17948428

fbshipit-source-id: 64599e459b95fe42bfbe1a19f54d0d6b1da90799