random taskset and log identifier (#384)
Summary:
Pull Request resolved: https://github.com/facebook/FAI-PEP/pull/384

Previously, we fix cpu core to 4. We have logic cpu core #4-#13 all isolated. In order to fully utilize the feature, we should randomly assign the job to a cpu core.

Also, we want to add a logging message with identifier so it will easy to find it when debugging.

Reviewed By: houseroad

Differential Revision: D18306936

fbshipit-source-id: 877e7e1c3574456bb9aab6a1f12b9c7138daa414