aibench mkstemp do not use job identifier
Summary:
In D20015348, one of the change is to add job identifier as part of prefix for `mkstemp`. From the runs in yesterday and today, I found that it will throw file or directory not found error randomly on our servers.

Will kindly revert the change for `mkstemp`. This is fine as we are cleaning it anyway after run.

Differential Revision: D20105809

fbshipit-source-id: e6833b3a005c2a4dcd058e4b0da785675ff5fd29