aibench mkstemp do not use job identifier (#399)
Summary:
Pull Request resolved: https://github.com/facebook/FAI-PEP/pull/399

In D20015348, one of the change is to add job identifier as part of prefix for `mkstemp`. From the runs in yesterday and today, I found that it will throw file or directory not found error randomly on our servers.

Will kindly revert the change for `mkstemp`. This is fine as we are cleaning it anyway after run.

Reviewed By: houseroad

Differential Revision: D20105809

fbshipit-source-id: 31c264234c8ce04179081e97bc97fc52738c6cdb