memory efficient way to calculate md5 for large file (#418)
Summary:
Pull Request resolved: https://github.com/facebook/FAI-PEP/pull/418

There were multi places in aibench to calculate md5sum of files.

For small files, this is ok but not ok if we are reading large files and hosts does not have enough memory to fit the whole file.

This diff change the way to load the file chunk by chunk, and only calculate md5sum once for large file per run.

Reviewed By: houseroad

Differential Revision: D21050879

fbshipit-source-id: e38e17841d3b15fb927eea2da81efa4d24e6039e