Add complete error logging when request data failed.
Summary:
Trying to debug error in https://fb.workplace.com/groups/1690415234371429/permalink/3443031369109798/.
The error happens when validate the result by check whether 'status' field is present and the value of 'status' is success, if not, then throw error.
But there are no error generated in the 'requestsJson' method, this means the result is being returned.
We need to verify whether the failure is coming from the real db post error with detailed error message or result format change.

Reviewed By: linbinyu

Differential Revision: D24968369

fbshipit-source-id: 6c6ef76cbd9744025f6eb69cc49df3f4e288ee6c