Device Monitoring
Summary: Adds device health monitoring to AIBench binary.  DeviceManager class will be instantiated by RunLab object and maintain the list of devices, as well as polling connections if lab platform is mobile (ios or android).  Disconnected devices will update the device status for the lab, and to the db.  Newly connected devices will be added to lab devices and db.

Reviewed By: kimishpatel

Differential Revision: D25738339

fbshipit-source-id: 817d81a4ec0733b7db4ac9f10548ac06aa3b5387