Fix iOS driver set error when providing device.
Summary: This change is required to fix an error where new devices detected by device monitor were passed in as a list to instantiate the platform.  This was causing a "cannot index on set" error from the removed line.  Platform should be instantiated by its hash, which is passed as the device argument.

Differential Revision: D27634546

fbshipit-source-id: 72571c2df3f1ba5ea2d34aa08ee09785dfbfa4f6