Fix iOS driver set error when providing device. (#442)
Summary:
Pull Request resolved: https://github.com/facebook/FAI-PEP/pull/442

This change is required to fix an error where new devices detected by device monitor were passed in as a list to instantiate the platform.  This was causing a "cannot index on set" error from the removed line.  Platform should be instantiated by its hash, which is passed as the device argument.

Reviewed By: axitkhurana

Differential Revision: D27634546

fbshipit-source-id: acd4af10432a704092e6ec9ce1c2f81098b75239