Add support for custom statistics, cleanup, and apply Black formatting.
Summary:
Add a new optional statistics field to the model's test .json to allow overriding the default statistics with a custom set.

To exercise this new feature, simply add an entry similar to the following examples under the tests section in the <model>.json file:
```
        "statistics": [ "p10","p90", "p50", "p95", "p5", "p11" ],
```
or
```
          "statistics": [
            "p10",
            "p90",
            "mean",
            "p95",
            "p5",
            "p99"
          ],
```

If you omit "p50", it will be appended to the list because it is required for some basic internal calculations.

The AIBench results web page was updated as a separate www change.

Reviewed By: MarkAndersonIX

Differential Revision: D30561031

fbshipit-source-id: 412c25e862e68edf90b47a6a980fea2a346a5aa9