Add support for custom statistics. (#454)
Summary:
Pull Request resolved: https://github.com/facebook/FAI-PEP/pull/454

Add a new optional statistics field to the model's test .json to allow overriding the default statistics with a custom set.
To exercise this new feature, simply add an entry similar to the following examples under the tests section in the <model>.json file:
        "statistics": [ "p10","p90", "p50", "p95", "p5", "p11" ],
or
          "statistics": [
            "p10",
            "p90",
            "mean",
            "p95",
            "p5",
            "p99"
          ],
If you omit "p50", it will be appended to the list because it is required for some basic internal calculations.

Reviewed By: axitkhurana

Differential Revision: D30921603

fbshipit-source-id: 8d85f1f90c1faee724820cc624b74bbe96bab23c