Include meta data when writing logs to user specified directory
Summary:
This change adds the ability to log the meta data (collected during the benchmark run) to the user specified directory.

This is specifically useful when we generate the flame-graphs using the `--force_profile`. The links to the flame-graph are now included in the meta data and stored with the logs.

Reviewed By: axitkhurana

Differential Revision: D31412313

fbshipit-source-id: e7e85af97dcbc22121509d83f6337ac7c2396f0c