Quiet gratuitous error logging for isRooted() check
Summary:
Quiet gratuitous error logging for isRooted() check.

We were still longing an "error" for the IsRooted devices even though that is 100% expected. Added the optional ignore_status=True flag to the calls to prevent that.

Also made sure we don't give the "Retrying 0 more times message" when not retrying.

Reviewed By: MarkAndersonIX

Differential Revision: D32711720

fbshipit-source-id: 685506a303a1fe518d61cadaf963228d2dcfc935