Quiet gratuitous error logging for isRooted() check (#465)
Summary:
Pull Request resolved: https://github.com/facebook/FAI-PEP/pull/465

Quiet gratuitous error logging for isRooted() check.

We were still longing an "error" for the IsRooted devices even though that is 100% expected. Added the optional ignore_status=True flag to the calls to prevent that.

Also made sure we don't give the "Retrying 0 more times message" when not retrying.

Reviewed By: MarkAndersonIX

Differential Revision: D32711720

fbshipit-source-id: 62a621135d42fd08180997835a8d37e0beb3c1f0