Add "report" option to degrade
Summary:
Add "report" option to degrade.

Can be used with or without a degrade.

Report also "works" on unrooted devices, since it is only reading,

Results may not be 100% "complete" on all devices, depending on what each device supports.

For example, not all devices support (the same set of) Governors.

Reviewed By: axitkhurana

Differential Revision: D33152644

fbshipit-source-id: 606c00df7da19a7337773c6784239e17eb186f6d