Add "report" option to degrade (#467)
Summary:
Pull Request resolved: https://github.com/facebook/FAI-PEP/pull/467

Add "report" option to degrade.

Can be used with or without a degrade.

Report also "works" on unrooted devices, since it is only reading,

Results may not be 100% "complete" on all devices, depending on what each device supports.

For example, not all devices support (the same set of) Governors.

Reviewed By: axitkhurana

Differential Revision: D33152644

fbshipit-source-id: fe8b342e2ac3fc10a6051ae7220ec59e14717992