AIBench: Memory profiling on Android 10+ devices using Perfetto
Summary:
Use perfetto (built into all Android devices from os 10 on) to generate a native heap profile for analysis.

This feature extends the existing SimplePerf cpu-profiling feature by adding a type (cpu vs memory) and optional options field to the config / AIBenchAPI.

Options may need to be specified to get useful results for a given benchmark type. Also, results are more interesting on devices running Android OS 12 or at least 11. It does not work at all on devices running less than Android OS 10.

For now, we substitute a perfetto_config file for the perfetto_report link instead of the intended flamegraph output link.

Output can be observed by downloading the perfetto_data link and importing it into https://ui.perfetto.dev.

__Caveats__:
1. Currently, the Perfetto Report link does not bring up the flamegraph directly. See workaround under test cases.
2. More tuning is needed on the default profile options based on the run size and duration.
3. User guidance needs to be documented for how to override the default options when necessary.

__AIBenchAPI changes__:

Amended the existing aibenchAPI profile argument to use a single bool || dict argument.

For example, for a test case below, here is the input:
```
profile={
          "profiler": "perfetto",
          "types": ["memory"],
          "options": {
            "shmem_size_bytes": "33554432",
            "sampling_interval_bytes": "2048",
            "buffer_size_kb": "262144",
            "buffer_size2_kb": "2048",
            "max_file_size_bytes": "200000000"
          }
        }
```
And here is what this translates to in the benchmark_config:
```
          "profiler": {
            "enabled": true,
            "profiler": "perfetto",
            "types": [
              "memory"
            ],
            "options": {
              "shmem_size_bytes": "33554432",
              "sampling_interval_bytes": "2048",
              "buffer_size_kb": "262144",
              "buffer_size2_kb": "2048",
              "max_file_size_bytes": "200000000"
            }
          }
```

In this case, the profiler flavor is redundant since if left unspecified, it will default to "perfetto" unless "cpu" (or nothing) is specified for types.

For simpleperf profiling, this argument can simply be
```
profile=True
```
or
```
profile={
          "types": ["cpu"]
}
```
or
```
profile={
          "profiler": "simpleperf"
}
```

Reviewed By: axitkhurana

Differential Revision: D31508439

fbshipit-source-id: bba4538de8a792bbb7ae5e11a23c1fd889d55dd3