Ensure profiling results filename is always unique
Summary:
Ensure profiling results filename is always unique.

It turns out we must not depend on the fact that we upload results to a unique folder.

This turns out to be useless if not redundant because once a result is download to (say) the users downloads folder, it will overwrite the local results of a previous run on the same device.

The device name is still very useful to tell which result you are looking for, but it is not enough.

Made the guid unconditional.

Differential Revision: D35365811

fbshipit-source-id: b67582f603f18699e88d9f15148e048786a6f20b