Add battery disconnect feature to perfetto battery runs
Summary:
Add battery disconnect / reconnect feature to perfetto battery runs.

This is a generic version which should work on many Android devices.

Still to come is to add a way to override the default mechanism for specific devices, if nece3ssary.

This approach will probably ONLY work on rooted devices since we need to write to file(s) under /sys/.

A follow-on diff will add a check for whether a device has a battery and give an error if you attempt to run battery profiling without a battery.

Reviewed By: axitkhurana

Differential Revision: D37129616

fbshipit-source-id: 09e2c700496b9acd21b36f9be2b3b0b6292fec3b