Add Counter and ODSCounter classes to AIBench (#511)
Summary:
Pull Request resolved: https://github.com/facebook/FAI-PEP/pull/511

Adds Counter class to OSS, and ODSCounter class to internal.  Counter can be used to bump counters with data given a key and a value as a float.  DeviceMonitor thread will now bump counters in ODS at entity "aibench_devices" for device connctivity based on the "live" value in the device dict.  Key is set to <hash>.connected.

Reviewed By: axitkhurana

Differential Revision: D37271565

fbshipit-source-id: 53448faca8fdbea6602c591b584bd31d72c4ce55