iOS Fix Multi-device Runs due to FileNotFound Error (#523)
Summary:
Pull Request resolved: https://github.com/facebook/FAI-PEP/pull/523

Where several devices or benchmarks are running at the time time on the same ios program, moving the filename to include .ipa extension can cause FileNotFound errors.  Replaces rename with shutil.copyfile and does a check for modified time to see whether it should copy the file.  This is a temporary fix and better download logic should be implemented longterm.

Reviewed By: axitkhurana

Differential Revision: D38891186

fbshipit-source-id: 77aaefd7580e2020338b5ba87cbdc420a642bfa8