Add extra AIBench env to the program run environment
Summary:
In the AIBench program, we need to know whether it is running as the control group or treatment group, so we are setting up this environment variable.

This feature is useful for the benchmark to determine things that should only be done in the treatment group, such as logging metrics to scuba.

Reviewed By: vmpuri

Differential Revision: D45839863

fbshipit-source-id: 8371c09306b84e9fcf55f72a3b7d98931b6f13c5