Thermal Monitoring and Shutdown Handling (#538)
Summary:
Pull Request resolved: https://github.com/facebook/FAI-PEP/pull/538

thermal shutdown handling for Android devices (i.e. small form factor devices which can exceed thermal limits running longer benchmarks)
Thermal monitoring can be enabled by adding a configuration with to the JSON file with the entry as the device's platform name.
The config should include the script to be pushed, thermal probe file location, and expression that evaluates to the thermal threshold.

    "PlatformName": {
        "script": "specifications/platform_scripts/thermal_monitoring/android_thermal_monitor.sh",
        "trip_temp_expr": "$(($(cat /device/path/to/trip_point_temp)-3000))",
        "temp_probe": " /device/path/to/temp_probe"
    }
Thermal monitoring is implemented using a context manager class that will push the script to the device and run it on an interval (default is 10.0 s) to check the thermal probe value and compare it to the trip_temp_expression.  If it is exceeded, the benchmark's pattern will be used to kill the program immediately, and logs will be appended to the benchmark run logs.

Reviewed By: xcheng16

Differential Revision: D38933376

fbshipit-source-id: 6cc7319090195c2f694ff017965d87f0903a4cbb