Refactor executorch framework handling
Summary:
Benchmarks under the Executorch framework function poorly since the handler is not in the correct place.

`runOnPlatform` is meant to be a generic function to run *any* command on the device. However, in this case, a handler for ET profiling data is called each time the this function is called that looks for a prof_result.bin file.

Therefore, if we have something in preprocess, the benchmark attempts to pull a file from the device (in this case, the platform is host, so the file is being pulled from the host to the host, a redundant operation) and typically fails because this file isn't (and shouldn't be) already present.

Reviewed By: tarun292

Differential Revision: D47145894

fbshipit-source-id: 3c8444cc35d25c69b9bb31a925b9ed1064799349