Remove unused import six.string_types from ios_driver.py
Summary:
## Instructions about RACER Diffs:
This diff is removing the unused imports from Code Quality Score from [Monetization codehub](https://fburl.com/quality/58ozwacb):
 - Accept and ship the diff (racer do not do auto land)
 - [**Recommended**]Commandeer and land this diff.

If you are happy with the changes, please accept and ship, or commandeer it if minor edits are needed.
If you are not happy with the changes, please comment on the diff with clear actions and send it back to the author. Racer will pick it up and re-generage.
If you really feel the Racer is not helping with this change (alas, some complex changes are hard for AI) feel free to abandon this diff.

This diff was pre-created by Racer AI agent for your convenience on top of T223416070. More details about the program see this [post](https://fb.workplace.com/groups/1177219810659335/permalink/1198503531864296/). For questions or suggestions please post in [Racer Users and Feedbacks](https://fb.workplace.com/groups/1112800313955334) group.
## Summary:
Removed unused import 'six.string_types' from ios_driver.py as identified by the F401 linting error.
 ---
> Generated by [RACER](https://www.internalfb.com/wiki/RACER_(Risk-Aware_Code_Editing_and_Refactoring)/), powered by [Confucius](https://www.internalfb.com/wiki/Confucius/Analect/Shared_Analects/Confucius_Code_Assist_(CCA)/)
[Session](https://www.internalfb.com/confucius?session_id=85456295-3849-11f0-a822-eae61486ffa4&tab=Chat), [Trace](https://www.internalfb.com/confucius?session_id=85456295-3849-11f0-a822-eae61486ffa4&tab=Trace)

Differential Revision: D74556512

fbshipit-source-id: 1055df50a451086018dc428d8b4fee5938a2baf8