Fix ios run end detector
Summary:
`self.util.listFiles()` returns a list of files(P2026824696). The if condition here checks if "BENCH_DONE" is one of the elements in `self.util.listFiles()`. But we want to check if it exists as a substring in any of the elements in the list. This diff fixes it.

https://fb.workplace.com/groups/1690415234371429/permalink/24936369469349344/

Reviewed By: marcinkwiatkowski

Differential Revision: D86445610

fbshipit-source-id: 0339bd104a8c0c020a24adc86eb7d56d73a0cd0a